+++ docs/.vuepress/components/jsonapi.vue
@@ -1,0 +1,21 @@
+<template>
+  <span class="j"><span class="g">{</span>json<span class="g">:</span>api<span class="g">}</span></span>
+</template>
+
+<style scoped>
+.g {
+  color: gray;
+  font-size: 1.2em;
+  line-height: normal;
+}
+
+.g:before, .g:after {
+  letter-spacing: 1.1em;
+}
+
+.j {
+  font-family: monospace;
+  font-weight: 700;
+  line-height: normal;
+}
+</style>