+++ src/api/ResourcefulApi.js
@@ -1,14 +1,12 @@
 import normalize from 'json-api-normalizer'
 
 import { Api } from './Api'
+import { ApiError, NotFoundApiError } from '../errors/ApiError'
 import { ModuleBuilder } from '../module/ModuleBuilder'
 import { ResourceProxy } from './ResourceProxy'
 import { deref, hasOwn } from '../shared/utils'
 import { Performance } from '../shared/Performance'
 
-/**
- * @class ResourcefulApi
- */
 export class ResourcefulApi extends Api {
   /**
    * Extends `Api::doRequest()` to handle some data preprocessing.
@@ -21,25 +19,78 @@ export class ResourcefulApi extends Api {
    * @param {String} url
    * @param {Object} params
    * @param {Object} data
-   * @param {Object} options
    */
-  async doRequest (method, url, params, data, options) {
+  async _doRequest (method, url, params, data) {
     if (data) {
       data = this.preprocessData(data)
     }
 
-    return super.doRequest(method, url, params, data, options)
-      .then((response) => {
-        return {
-          data: normalize(response.data),
-          meta: response.data.meta,
-          status: response.status
+    return super._doRequest(method, url, params, data)
+      .then(async (response) => {
+        if (this._shouldDecodeResponseJson(response.status)) {
+          const json = await this._decodeResponseJson(response)
+
+          return this._parseResponse(response.status, json)
         }
+
+        return this._createDatalessResponse(response)
       })
   }
 
+  _shouldDecodeResponseJson (status) {
+    return [200, 201, 404, 409].includes(status)
+  }
+
+  async _decodeResponseJson (response) {
+    try {
+      return await response.json()
+    } catch (e) {
+      if (response.status === 404) {
+        throw new NotFoundApiError('Resource not found')
+      }
+
+      throw new ApiError('Failed to decode response json')
+    }
+  }
+
+  _parseResponse (status, json) {
+    if (!(hasOwn(json, 'data') || hasOwn(json, 'errors'))) {
+      throw new ApiError('Response object must have either a `data` or an `errors` property.')
+    }
+
+    const parsedResponse = {
+      meta: json.meta ? json.meta : {},
+      links: json.links ? json.links : {},
+      status: status
+    }
+
+    if (json.data) {
+      parsedResponse.data = normalize(json)
+    }
+
+    if (json.errors) {
+      parsedResponse.errors = json.errors
+
+      switch (status) {
+        case 404:
+          throw new NotFoundApiError('Resource not found but received error info', json.errors)
+      }
+    }
+
+    return parsedResponse
+  }
+
+  _createDatalessResponse (response) {
+    return {
+      data: null,
+      meta: null,
+      links: null,
+      status: response.status
+    }
+  }
+
   /**
-   * convert RessourceTypes to uppercase
+   * convert ResourceTypes to uppercase
    * to follow the json:api spects even if the incoming data is not correct
    *
    * this is just a safety net
@@ -102,7 +153,7 @@ export class ResourcefulApi extends Api {
   /**
    * Prepare the routable requests
    *
-   * @param {route.Router} router
+   * @param {Router} router
    */
   setupResourcefulRequests (router) {
     this.router = router
@@ -112,13 +163,11 @@ export class ResourcefulApi extends Api {
     const routes = router.getRoutes()
     this.registerableModules = {}
 
-    for (const routeName in routes) {
-      if (hasOwn(routes, routeName)) {
-        const methods = routes[routeName]
+    for (const routeName of Object.keys(routes)) {
+      const methods = routes[routeName]
 
-        this.registerResourceMethods(routeName, methods)
-        this.registerableModules[routeName] = methods
-      }
+      this.registerResourceMethods(routeName, methods)
+      this.registerableModules[routeName] = methods
     }
 
     Performance.mark('api_setup_routing_end')
@@ -171,7 +220,7 @@ export class ResourcefulApi extends Api {
       return
     }
 
-    const moduleBuilder = new ModuleBuilder(this.store, this, moduleName, methods)
+    const moduleBuilder = new ModuleBuilder(this, moduleName, methods)
     const module = moduleBuilder.build()
     if (moduleName) {
       this.store.registerModule(moduleName, module)
@@ -190,16 +239,14 @@ export class ResourcefulApi extends Api {
 
     const relationsToBeAdded = []
 
-    for (const methodName in methods) {
-      if (hasOwn(methods, methodName)) {
-        if (this.isRelationMethodName(methodName)) {
-          relationsToBeAdded.push(methodName)
+    for (const methodName of Object.keys(methods)) {
+      if (this.isRelationMethodName(methodName)) {
+        relationsToBeAdded.push(methodName)
 
-          continue
-        }
-
-        this[routeName].addRoute(methods[methodName])
+        continue
       }
+
+      this[routeName].addRoute(methods[methodName])
     }
 
     relationsToBeAdded.forEach(relationToBeAdded => {