+++ src/shared/utils.js
@@ -99,20 +99,5 @@ export function checkConfigProperty (config, property, isRequiredProp = true) {
  * @returns {boolean}
  */
 export function validateCallbackFn (fn) {
-  return fn.constructor === Function
-}
-
-/**
- * Make sure all elements of a passed array are callable functions
- *
- * @param {Function[]} callbacks
- */
-export function validateCallbackFns (callbacks) {
-  if (typeof callbacks === 'undefined' ||
-    callbacks.constructor !== Array ||
-    callbacks.reduce((carry, cb) => validateCallbackFn(cb) && carry, true) === false) {
-    throw new Error('You must pass an array of valid callback functions to this method')
-  }
-
-  return callbacks
+  return fn instanceof Function
 }